requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
tzdata==2024.1
Pillow==10.3.0
python-decouple
djangorestframework
